hermod-python/src/requirements.txt
soundfile
paho-mqtt
deepspeech
pyaudio
webrtcvad
numpy
scipy
paho-mqtt
pvporcupine
pyyaml
rasa
hbmqtt
pyyaml
rasa-sdk
flask-dance
gunicorn
blinker
asyncio_mqtt
python-dotenv
pymediawiki
word2number
wikidata
sounddevice
wptools
google-cloud-speech
